Basic outline of purchase method, which is used by the user when a purchase happens, is completed.
The method itemPurchased is completed. This puts the item and quantity
of the item that have been bought into a list.

The method anotherPurchase is completed. This allows the user to input
another item of a different type.

The pricing method has been started but needs database features in order
to be completed.

The receipt method has been started but needs database in order to
continue.