Added a main for testing Transaction.
Added in comments so that someone else can understand the code.

Changed the purchase method so that it effectively uses the other
methods and works around common bugs that was making the solution
more complicated.

Changed from using lists of lists, to a list of strings, where it splits
the strings based on commas that are between the variables that are
needed.

Finished off receipt so that it prints out the item quantity and price.